PROYECTO CON Polices y gates

+++ app/Http/Controllers/JugadorController.php
@@ -53,7 +53,9 @@ class JugadorController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function show(Jugador $jugadores)
-    {
+    { 
+        //dd($jugadores);
+        $this->authorize('jugadorp',$jugadores);
         return view('jugadores.jugadoresShow',compact('jugadores'));
     }
 
@@ -64,7 +66,8 @@ class JugadorController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function edit(Jugador $jugadores)
-    {
+    {  
+        $this->authorize('jugadorp',$jugadores);
         return view('jugadores.jugadoresForm',compact('jugadores'));
     }
 
@@ -76,7 +79,10 @@ class JugadorController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function update(Request $request, Jugador $jugadores)
-    {
+    {  
+
+        $this->authorize('jugadorp',$jugadores);
+        
         $this->validate($request,['nombre' => 'required|string|min:4|max:8', 
         'edad' => 'required|integer|between:10,20'] ); 
         
@@ -98,7 +104,8 @@ class JugadorController extends Controller
      * @return \Illuminate\Http\Response
      */
     public function destroy(Jugador $jugadores)
-    {
+    { 
+        $this->authorize('view',$jugadores);
         $jugadores->delete();  
         session()->flash('statusj','Destruido LOGICO realizado');
         return redirect()->route('jugadores.index');
